Restructure nyxx_extensions
