+++ lib/nyxx_extensions.dart
@@ -1,15 +1,30 @@
+/// Extensions and additional utilities for working with [nyxx](https://pub.dev/packages/nyxx).
 library nyxx_extensions;
 
-export 'src/channel.dart';
-export 'src/date_time.dart';
-export 'src/embed_builder.dart';
-export 'src/emoji.dart';
-export 'src/guild.dart';
-export 'src/message.dart';
-export 'src/pagination.dart';
-export 'src/role.dart';
-export 'src/sanitizer.dart';
-export 'src/user.dart';
+export 'src/utils/emoji.dart';
+export 'src/utils/endpoint_paginator.dart' hide streamPaginatedEndpoint;
 export 'src/utils/formatters.dart';
-export 'src/permissions.dart';
-export 'src/endpoint_paginator.dart';
+export 'src/utils/pagination.dart';
+export 'src/utils/permissions.dart';
+export 'src/utils/sanitizer.dart';
+
+export 'src/extensions/channel.dart';
+export 'src/extensions/client.dart';
+export 'src/extensions/date_time.dart';
+export 'src/extensions/embed.dart';
+export 'src/extensions/emoji.dart';
+export 'src/extensions/guild.dart';
+export 'src/extensions/managers/audit_log_manager.dart';
+export 'src/extensions/managers/channel_manager.dart';
+export 'src/extensions/managers/entitlement_manager.dart';
+export 'src/extensions/managers/guild_manager.dart';
+export 'src/extensions/managers/member_manager.dart';
+export 'src/extensions/managers/message_manager.dart';
+export 'src/extensions/managers/scheduled_event_manager.dart';
+export 'src/extensions/managers/user_manager.dart';
+export 'src/extensions/member.dart';
+export 'src/extensions/message.dart';
+export 'src/extensions/role.dart';
+export 'src/extensions/scheduled_event.dart';
+export 'src/extensions/snowflake_entity.dart';
+export 'src/extensions/user.dart';
