+++ lib/src/extensions/client.dart
@@ -1,0 +1,13 @@
+import 'package:nyxx/nyxx.dart';
+import 'package:nyxx_extensions/src/utils/emoji.dart';
+
+/// Extensions on [NyxxRest].
+extension NyxxRestExtensions on NyxxRest {
+  /// List all the text emoji available to this client.
+  Future<List<TextEmoji>> getTextEmojis() async => (await getEmojiDefinitions())
+      .map((definition) => TextEmoji(id: Snowflake.zero, manager: guilds[Snowflake.zero].emojis, name: definition.surrogates))
+      .toList();
+
+  /// Get a text emoji by name.
+  TextEmoji getTextEmoji(String name) => TextEmoji(id: Snowflake.zero, manager: guilds[Snowflake.zero].emojis, name: name);
+}
