+++ lib/src/extensions/message.dart
@@ -1,7 +1,8 @@
 import 'package:nyxx/nyxx.dart';
-import 'package:nyxx_extensions/src/channel.dart';
+import 'package:nyxx_extensions/src/extensions/channel.dart';
 
-extension MessageUtils on Message {
+/// Extensions on [Message]s.
+extension MessageExtensions on Message {
   /// A URL clients can visit to navigate to this message.
   Future<Uri> get url async => Uri.https(manager.client.apiOptions.host, '${(await channel.get()).url.path}/$id');
