+++ lib/src/extensions/role.dart
@@ -1,7 +1,8 @@
 import 'package:nyxx/nyxx.dart';
 import 'package:nyxx_extensions/src/utils/formatters.dart';
 
-extension RoleExtension on PartialRole {
+/// Extensions on [PartialRole]s.
+extension PartialRoleExtensions on PartialRole {
   /// A mention of this role.
   String get mention {
     if (id == manager.guildId) {
